chore: move terminalview into a subfolder
