+++ src/controller/controller.h
@@ -6,7 +6,7 @@
 #include "../model/object/entity/individual/fred/fred.h"
 #include "../model/object/entity/individual/piercy/piercy.h"
 #include "../model/random/random.h"
-#include "../view/terminalview.h"
+#include "../view/terminalview/terminalview.h"
 #include "../model/world/world.h"
 
 #include <cmath>
