+++ src/view/terminalview/terminalview.cpp
