+++ src/view/terminalview/terminalview.h
@@ -2,7 +2,7 @@
 #ifndef EVE_TERMINALVIEW_H
 #define EVE_TERMINALVIEW_H
 
-#include "../model/world/world.h"
+#include "../../model/world/world.h"
 
 class TerminalView {
 private:
