+++ test/view/terminalview/terminalview.cpp
@@ -1,6 +1,6 @@
 /*
-#include "../../src/view/terminalview.h"
-#include "../../src/model/object/entity/individual/fred/fred.h"
+#include "../../../src/view/terminalview/terminalview.h"
+#include "../../../src/model/object/entity/individual/fred/fred.h"
 #include <catch2/catch.hpp>
 #include <exception>
